requirements.txt
azure-storage-blob
pandas
tabulate
